LPD-82334 Reorganize Object Action tests into existing files
