+++ modules/test/playwright/tests/object-web/client-extension/objectClientExtension.spec.ts
@@ -342,3 +342,35 @@ test('Can trigger object validation as a client extension', async ({
 		page.locator('.cell-name').getByText('Valid Name')
 	).toBeVisible();
 });
+
+test(
+	'LPD-78504 Can trigger action with unmodifiable system object definition using client extension',
+	{tag: '@LPD-78504'},
+	async ({apiHelpers, editObjectActionPage, page, viewObjectActionsPage}) => {
+		const notificationTemplate =
+			await apiHelpers.notification.postRandomNotificationTemplate(
+				'notificationTemplate' + getRandomInt()
+			);
+
+		apiHelpers.data.push({
+			id: notificationTemplate.id,
+			type: 'notificationTemplate',
+		});
+
+		// Navigate to a system object's actions page (User is an unmodifiable system object)
+
+		await viewObjectActionsPage.goto('User');
+
+		// Add a new action using notification type on the system object
+
+		await editObjectActionPage.addNewAction({
+			notificationTemplateName: notificationTemplate.name,
+			thenOption: 'Notification',
+			whenOption: 'On After Add',
+		});
+
+		// Verify the action was created by checking it appears in the actions list
+
+		await expect(page.getByText('On After Add')).toBeVisible();
+	}
+);
